// repositorio/SIMTR/4-implementacao/simtr-web/src/app/cruds/model/crud-produto.ts
export class CrudProduto
{
    identificador_produto?: number;
    nome_produto?: string;
    operacao_produto?: any;
    modalidade_produto?: any;
    indicador_contratacao_conjunta?: boolean;
    indicador_dossie_digital?: boolean;
    indicador_pesquisa_cadin?: boolean;
    indicador_pesquisa_ccf?: boolean;
    indicador_pesquisa_receita?: boolean;
    indicador_pesquisa_scpc?: boolean;
    indicador_pesquisa_serasa?: boolean;
    indicador_pesquisa_sicow?: boolean;
    indicador_pesquisa_sinad?: boolean;
    indicador_tipo_pessoa?: string;
    codigo_produto_portal_empreendedor?: number;
}